pyproject.toml
[tool.poetry]
name = "cdk-s3-cloudfront-example"
version = "0.1.0"
description = ""
authors = ["William Chou <williamchou@kklab.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.140.0"
constructs = "^10.3.0"
ruff = "^0.4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
